Move file targets into separate folders
